Exclude `jackson-module-kotlin` to avoid CP issue
The `Jackson2JsonObjectMapper` tries to load any available modules for
the Jackson `ObjectMapper` via `findAndRegisterModules()`.

But if we have a module but don't have a target library, we end up with
the `ClassNotFoundException`.
In our case we fail with the Kotlin because Spring Boot provides
`com.fasterxml.jackson.module:jackson-module-kotlin` dependency and
leaves the rest of required libs for the target application.

* Exclude `com.fasterxml.jackson.module:jackson-module-kotlin`
from all the dependencies.
* Spring Integration `Jackson2JsonObjectMapper` should be revised to use
`Jackson2ObjectMapperBuilder` - https://jira.spring.io/browse/INT-4318