Use `cacheConsumers=false` for  CachingJmsCF
https://build.spring.io/browse/INTSAMPLES-NIGHTLY-2256

The `AggregatorDemoTest` fails sporadically with no second part from
reply.
That might be as a result of the cached consumers in the
`CachingConnectionFactory`, so several `<int-jms:inbound-gateway>`
for the topic using the same connectionFactory essentially end up
using the same consumer sequentially.
And Broker considers them as a single consumer, therefore only one
message is sent.