GH-361: Add `timeout` and `scheduling.pool.size` for Cafe DSL demo
Fixes: #361

The default request timeout in Spring Integration is `30 seconds`
to not block the thread forever.
However, the logic in the Cafe Demo is longer even that one minute
to wait for all the drinks to be ready to free an order queue.

* Add `spring.integration.endpoint.default-timeout=-1` to wait on the customer request forever for the logic of the application
* Add `spring.task.scheduling.pool.size=10` to avoid deadlock when all the scheduled tasks compete for the same thread by default
* Upgrade to latest Gradle
* Upgrade to the latest Spring Boot

Related to: https://github.com/spring-projects/spring-boot/issues/41477